mgr/dashboard_v2: Renamed/moved RESTResource to RESTController
